requirements.txt
# requirements.txt (2025年7月版 互換性OK)
fastapi
uvicorn
line-bot-sdk
python-dotenv
requests
pandas
faiss-cpu
# LangChain関連は最新版に統一
langchain
langchain-core
langchain-community
langchain-google-genai
google-generativeai
# Web検索API関連
google-search-results
langchain-google-community